feat: rename eg.Component to eg-component
