+++ README.md
@@ -1,1 +1,64 @@
-#eg.Component
+#egjs-component
+
+## requirement
+
+```
+// dependencies
+npm install
+```
+
+## build
+
+```
+npm run build //or webpack
+```
+
+## test
+
+```
+npm run test //or karma start
+```
+
+## documentation
+
+```
+npm run jsdoc //or jsdoc -c jsdoc.json
+```
+
+## dev server
+
+```
+npm run dev-server //or webpack-dev-server
+```
+
+checkout http://localhost:8080/demo/
+
+## directories
+
+```
+egjs-component
+  ├─ dist/
+  │   └─ component.js
+  │   └─ comopnent.min.js
+  │   └─ component.js.map
+  │   └─ component.min.js.map
+  │   └─ component.test.js
+  │   └─ component.min.js.map
+  ├─ src/
+  │   └─ index.js
+  ├─ doc/
+  ├─ test/
+  │   └─ index.js
+  ├─ .editconfig
+  ├─ .gitignore
+  ├─ .npmignore
+  ├─ bower.json
+  ├─ karma.conf.js
+  ├─ package.json
+  ├─ README.md
+  └─ webpack.config.js
+```
+
+## License
+
+MIT © NAVER Corp.
